Add ability to split partitions into unconnected subgraphs
This allows the merging to be more efficient - otherwise the default
partitions might have nodes combined that prevent more merging from
occuring than could otherwise be done.  Also fix a bug in the merging
algorithm where when we sorted the possible merges based on cancellation
numbers, we didn't reindex the ids identifying the merges canceled by
each merge with the new sorted indices.